requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiosignal==1.3.2
altair==5.5.0
anndata==0.10.9
appdirs==1.4.4
arboreto==0.1.6
array_api_compat==1.11.2
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
biopython==1.85
bokeh==3.4.3
boltons==25.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
cloudpickle==3.1.1
ConfigArgParse==1.7.1
connection-pool==0.0.3
contourpy==1.3.0
cramjam==2.10.0
ctxcore==0.2.0
cycler==0.12.1
cytoolz==1.0.1
dask==2024.8.0
dask-expr==1.1.10
datrie==0.8.2
dill==0.4.0
distributed==2024.8.0
docutils==0.21.2
dpath==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastjsonschema==2.21.1
fastparquet==2024.11.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.1
frozendict==2.4.6
frozenlist==1.6.0
fsspec==2025.5.1
gast==0.6.0
GEOparse==2.0.4
get-annotations==0.1.2
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
grpcio==1.71.0
gseapy==1.1.8
h5py==3.13.0
HTSeq==2.0.9
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
interlap==0.2.7
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_core==5.8.1
keras==3.10.0
kiwisolver==1.4.7
legacy-api-wrap==1.4.1
libclang==18.1.1
lightgbm==4.6.0
llvmlite==0.43.0
locket==1.0.0
loompy==3.0.8
lz4==4.4.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
multiprocessing_on_dill==3.5.0a4
namex==0.1.0
narwhals==1.41.0
natsort==8.4.0
nbformat==5.10.4
ncls==0.0.68
networkx==3.2.1
numba==0.60.0
numexpr==2.10.2
numpy==2.0.2
numpy-groupies==0.11.3
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pandas==2.2.3
partd==1.4.2
patsy==1.0.1
pillow==11.2.1
plac==1.4.5
platformdirs==4.3.8
plotly==6.1.2
propcache==0.3.1
protobuf==5.29.5
psutil==7.0.0
PuLP==3.2.1
pyarrow==20.0.0
pyarrow-hotfix==0.7
pybedtools==0.12.0
pyBigWig==0.3.24
PyBioMed==1.0
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.3
pyranges==0.1.4
pysam==0.23.1
pyscenic==0.12.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
reretry==0.11.8
rich==14.0.0
rpds-py==0.25.1
scanpy==1.10.3
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
session_info==1.0.1
shap==0.47.2
six==1.17.0
slicer==0.0.8
smart-open==7.1.0
smmap==5.0.2
snakemake==7.32.4
sorted_nearest==0.0.39
sortedcontainers==2.4.0
statsmodels==0.14.4
stdlib-list==0.11.1
stopit==1.1.2
sympy==1.14.0
tabulate==0.9.0
tblib==3.1.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
throttler==1.2.2
toolz==1.0.0
toposort==1.10
torch==2.7.0
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
umap-learn==0.5.7
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==2.1.4
xyzservices==2025.4.0
yarl==1.20.0
yte==1.7.0
zict==3.0.0
zipp==3.22.0
